+++ package.json
@@ -24,7 +24,11 @@
     "babel-jest": "^26.0.1",
     "eslint": "^6.8.0",
     "eslint-config-plazide": "^0.1.6",
+    "faunadb": "^3.0.1",
     "jest": "^26.0.1"
+  },
+  "peerDependencies": {
+    "faunadb": "^3.0.0"
   },
   "directories": {
     "lib": "lib"
